Split SurfaceGroupModel into its own file
SurfaceModel and SurfaceGroupModel are related but distinct:
SurfaceModel is the state for a single surface, SurfaceGroupModel
manages multiple surfaces and forwards events between them. Keeping
them in separate files makes each easier to find and aligns with the
one-class-per-file pattern used by the other models (ComponentModel,
DataModel).
