+++ packages/a2ui_core/lib/src/processing/processor.dart
@@ -8,6 +8,7 @@ import '../common/errors.dart';
 import '../core/catalog.dart';
 import '../core/component_model.dart';
 import '../core/messages.dart';
+import '../core/surface_group_model.dart';
 import '../core/surface_model.dart';
 
 /// The central processor for A2UI messages.
